multiplayer implementation

+++ src/players/Entity.java
@@ -4,14 +4,14 @@ import java.awt.*;
 
 public abstract class Entity {
 
-    protected int x;
-    protected int y;
+    protected double x;
+    protected double y;
     protected int width;
     protected int height;
     protected Color color;
 
-    public void render(Graphics g){
+    public void render(Graphics g) {
         g.setColor(color);
-        g.fillRect(x,y,width,height);
+        g.fillRect((int) x, (int) y, width, height);
     }
 }
